# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FisioGest Pro API - Clinical management platform for physiotherapy, aesthetics and pilates
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      security:
        - BearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [agendado, confirmado, concluido, cancelado, faltou]
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentWithDetails"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
        "409":
          description: Time slot conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
    put:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentRequest"
      responses:
        "200":
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel/delete an appointment
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /appointments/{id}/complete:
    post:
      operationId: completeAppointment
      tags: [appointments]
      summary: Mark appointment as completed and generate financial record
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
...
  /financial/dashboard:
    get:
      operationId: getFinancialDashboard
      tags: [financial]
      summary: Get financial dashboard data
      security:
        - BearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Financial dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialDashboard"
  /financial/records:
    get:
      operationId: listFinancialRecords
      tags: [financial]
      summary: List financial records
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [receita, despesa]
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Financial records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinancialRecord"
    post:
      operationId: createFinancialRecord
      tags: [financial]
      summary: Create a financial record (expense)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFinancialRecordRequest"
      responses:
        "201":
          description: Record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialRecord"
...
  /dashboard:
    get:
      operationId: getDashboard
      tags: [reports]
      summary: Main dashboard overview
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [admin, profissional, recepcionista]
          default: profissional
      required:
        - name
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
    Patient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cpf:
          type: string
        birthDate:
          type: string
          format: date
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        profession:
          type: string
        emergencyContact:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - cpf
        - phone
        - createdAt
...
    Procedure:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [fisioterapia, estetica, pilates]
        durationMinutes:
          type: integer
        price:
          type: number
        cost:
          type: number
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - durationMinutes
        - price
        - createdAt
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        procedureId:
          type: integer
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [agendado, confirmado, concluido, cancelado, faltou]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - patientId
        - procedureId
        - date
        - startTime
        - endTime
        - status
        - createdAt
    AppointmentWithDetails:
      allOf:
        - $ref: "#/components/schemas/Appointment"
        - type: object
          properties:
            patient:
              $ref: "#/components/schemas/Patient"
            procedure:
              $ref: "#/components/schemas/Procedure"
    CreateAppointmentRequest:
      type: object
      properties:
        patientId:
          type: integer
        procedureId:
          type: integer
        date:
          type: string
          format: date
        startTime:
          type: string
        notes:
          type: string
      required:
        - patientId
        - procedureId
        - date
        - startTime
    UpdateAppointmentRequest:
      type: object
      properties:
        patientId:
          type: integer
        procedureId:
          type: integer
        date:
          type: string
          format: date
        startTime:
          type: string
        status:
          type: string
          enum: [agendado, confirmado, concluido, cancelado, faltou]
        notes:
          type: string
...
    FinancialRecord:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [receita, despesa]
        amount:
          type: number
        description:
          type: string
        category:
          type: string
        appointmentId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - amount
        - description
        - createdAt
    CreateFinancialRecordRequest:
      type: object
      properties:
        type:
          type: string
          enum: [receita, despesa]
          default: despesa
        amount:
          type: number
        description:
          type: string
        category:
          type: string
      required:
        - amount
        - description
    FinancialDashboard:
      type: object
      properties:
        monthlyRevenue:
          type: number
        monthlyExpenses:
          type: number
        monthlyProfit:
          type: number
        averageTicket:
          type: number
        totalAppointments:
          type: integer
        completedAppointments:
          type: integer
        topProcedure:
          type: string
        revenueByCategory:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              revenue:
                type: number
            required:
              - category
              - revenue
      required:
        - monthlyRevenue
        - monthlyExpenses
        - monthlyProfit
        - averageTicket
        - totalAppointments
        - completedAppointments
    DashboardData:
      type: object
      properties:
        todayAppointments:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentWithDetails"
        upcomingAppointments:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentWithDetails"
        monthlyRevenue:
          type: number
        totalPatients:
          type: integer
        todayTotal:
          type: integer
        occupationRate:
          type: number
      required:
        - todayAppointments
        - upcomingAppointments
        - monthlyRevenue
        - totalPatients
        - todayTotal
        - occupationRate
...