# lib/api-zod/src/generated/types/createFinancialRecordRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FisioGest Pro API - Clinical management platform for physiotherapy, aesthetics and pilates
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFinancialRecordRequestType } from "./createFinancialRecordRequestType";

export interface CreateFinancialRecordRequest {
  type?: CreateFinancialRecordRequestType;
  amount: number;
  description: string;
  category?: string;
}