# lib/api-zod/src/generated/types/financialRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FisioGest Pro API - Clinical management platform for physiotherapy, aesthetics and pilates
 * OpenAPI spec version: 0.1.0
 */
import type { FinancialRecordType } from "./financialRecordType";

export interface FinancialRecord {
  id: number;
  type: FinancialRecordType;
  amount: number;
  description: string;
  category?: string;
  appointmentId?: number;
  createdAt: Date;
}