# lib/api-zod/src/generated/types/procedure.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FisioGest Pro API - Clinical management platform for physiotherapy, aesthetics and pilates
 * OpenAPI spec version: 0.1.0
 */
import type { ProcedureCategory } from "./procedureCategory";

export interface Procedure {
  id: number;
  name: string;
  category: ProcedureCategory;
  durationMinutes: number;
  price: number;
  cost?: number;
  description?: string;
  createdAt: Date;
}